Adding and removing LEDs
This seems to work for one LED, yet the leds-gpio driver doesn't
implement a release hook as nobody has probably used it this way.
I'll have to patch the leds-gpio driver to allow for disconnecting
itself.

There was some strange segfault when trying to setup a multipin
socket for more LEDs but it may have been due to that Oops before.

More testing is definitely needed.